feat: remove the unreachable Madagascar from the map material
Madagascar cannot be reached in the game (the channel blocks, no ferry
serves it), so it leaves the map on user request: its world-trim seed
and its coastline polygon are removed, which lets the load-time trim
stamp the island to deep ocean and takes it off the exploration map's
ink sketch in one step.

Its removal surfaced two ghost traces in the water and both are closed
generally: shallow sea in the trimmed Madagascar box and northeast of
the Red Sea boundary is deepened up to 800 m (the steep old coast
slopes drew a pale outline at the former 150 m cap, which remains the
threshold only for the shelf dilation next to kept shores), and the
water's deep tone now saturates by ~1600 m so the -3000 m stamps no
longer stand out as darker silhouettes against shallower ocean basins.

design.md §3.1 now lists the reachable islands (Zanzibar, Pemba, Bioko)
and records Madagascar as trimmed; CLAUDE.md pt. 4, the README and the
redSea/world tests follow.

+++ src/render/water.ts
@@ -125,7 +125,10 @@ export function createWaterMaterial(): WaterMaterialHandle {
   const deep = color('#0b2f4e')
   let col = mix(riverTone, shallow, smoothstep(float(0), float(25), depthM))
   col = mix(col, mid, smoothstep(float(30), float(400), depthM))
-  col = mix(col, deep, smoothstep(float(500), float(3000), depthM))
+  // Fully deep by ~1600 m: abyssal plains, trimmed-land stamps (-3000 m,
+  // world/redSea.ts) and the outside-bbox mask all share one deep tone, so
+  // no trimmed silhouette stands out against a shallower ocean basin.
+  col = mix(col, deep, smoothstep(float(500), float(1600), depthM))
   // Gentle ripple modulation keeps large surfaces alive.
   const ripple = mx_fractal_noise_float(
     vec3(wp.x.mul(0.22).add(time.mul(0.12)), wp.y.mul(0.22), time.mul(0.07)),
